refactor: añadir manejo de excepciones

+++ sistemaGestionBiblioteca/src/main/java/app/biblioteca/sistemagestionbiblioteca/services/impl/LibroServiceImpl.java
@@ -1,5 +1,6 @@
 package app.biblioteca.sistemagestionbiblioteca.services.impl;
 
+import app.biblioteca.sistemagestionbiblioteca.exceptions.LibroNotFoundException;
 import app.biblioteca.sistemagestionbiblioteca.models.Libro;
 import app.biblioteca.sistemagestionbiblioteca.repositories.LibroRepository;
 import app.biblioteca.sistemagestionbiblioteca.services.LibroService;
@@ -18,7 +19,7 @@ public class LibroServiceImpl implements LibroService {
     @Override
     public Libro buscarPorIsbn(String isbn) {
         return libroRepository.findByIsbn(isbn)
-                .orElseThrow(() -> new RuntimeException("Libro no encontrado: " + isbn));
+                .orElseThrow(() -> new LibroNotFoundException(-1L));
     }
 
     @Override
@@ -33,13 +34,16 @@ public class LibroServiceImpl implements LibroService {
 
     @Override
     public void eliminar(Long id) {
+        if (!libroRepository.existsById(id)) {
+            throw new LibroNotFoundException(id);
+        }
         libroRepository.deleteById(id);
     }
 
     @Override
     public Libro actualizar(Long id, Libro libro) {
         if (!libroRepository.existsById(id)) {
-            throw new RuntimeException("Libro no encontrado con ID: " + id);
+            throw new LibroNotFoundException(id);
         }
         libro.setId(id);
         return libroRepository.save(libro);
